Check convergence of multiple MCMC chains

Uses Gelman multivariate Gelman-Rubin diagnostic to check if multiple MCMC chains have converged

Usage

check.convergence(jags_out, threshold = 1.1, verbose = TRUE, ...)

Arguments

jags_out

mcmc.list object (from coda package) containing samples from MCMC chains.

threshold

Gelman-Rubin diagnostic parameter threshold. Default = 1.1

verbose

If TRUE, print convergence result. Default = TRUE

...

Additional arguments to gelman.diag (from coda package)

Value

List length 3 containing the following: * convergence: Logical. Whether or not convergence was achieved. * diagnostics: Numerical value of Gelman-Rubin diagnostics for each parameter and multivariate diagnostic * error: Logical. Whether or not an error occured in the Gelman-Rubin calculation.